requirements.txt
cycler==0.10.0
graphviz==0.10.1
kiwisolver==1.0.1
matplotlib==2.2.3
nltk==3.4
numpy==1.15.4
opencv-python==4.0.0.21
paddlepaddle==1.2.0
Pillow==5.4.1
protobuf==3.1.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
rarfile==3.0
recordio==0.1.7
requests==2.9.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
